Finished all refactor for routes
